change getMsSinceEpoch to return type long long (I think this worked since it looks right on my Windows machine now)

+++ include/shared/utilities/time.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-long getMsSinceEpoch();
+long long getMsSinceEpoch();
