+++ src/shared/utilities/time.cpp
@@ -2,7 +2,7 @@
 
 #include <chrono>
 
-long getMsSinceEpoch() {
+long long getMsSinceEpoch() {
     return std::chrono::duration_cast<std::chrono::milliseconds>(
         std::chrono::system_clock::now().time_since_epoch()
     ).count();
